# Conflicts:
#	dashboard/db/schema.rb

+++ dashboard/app/helpers/users_helper.rb
@@ -11,15 +11,28 @@ module UsersHelper
       uid = session['clever_takeover_id']
       # TODO: validate that we're not destroying an active account?
       existing_clever_account = User.where(uid: uid).first
+
+      # Move over sections that students follow
+      if user.student?
+        Follower.where(student_user_id: existing_clever_account.id).each do |follower|
+          follower.update(student_user_id: user.id)
+        end
+      end
+
       existing_clever_account.destroy! if existing_clever_account
       user.provider = 'clever'
       user.uid = uid
       user.oauth_token = session['clever_takeover_token']
       user.save
-      session.delete('clever_link_flag')
-      session.delete('clever_takeover_id')
-      session.delete('clever_takeover_token')
+      clear_clever_session_variables
     end
+  end
+
+  def clear_clever_session_variables
+    return if session.empty?
+    session.delete('clever_link_flag')
+    session.delete('clever_takeover_id')
+    session.delete('clever_takeover_token')
   end
 
   # Summarize a user and his or her progress progress within a certain script.
